pyproject.toml
[project]
name = "openai-agents-sdk-mcp-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastmcp>=2.11.3",
    "openai-agents>=0.2.9",
    "playwright>=1.41.0",
]

[dependency-groups]
dev = [
    "anyio>=4.10.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
